🔥 앱 boilercode 리셋
폴더 안에 있는 내용물을 리셋. images폴더에 이모지, 기본 사진들 추가
