+++ oz-coding-school-app/app-example/(tabs)/explore.tsx
