+++ oz-coding-school-app/app-example/(tabs)/index.tsx
